Move persistence logic to its own file
